requirement.txt
streamlit
PyPDF2
python-dotenv
openai
langchain